+++ src/ts/main.ts
@@ -11,6 +11,7 @@ export { PlayerUtils } from './utils/PlayerUtils';
 export { UIUtils } from './utils/UIUtils';
 export { BrowserUtils } from './utils/BrowserUtils';
 export { StorageUtils } from './utils/StorageUtils';
+export { UIPreferences } from './utils/UIPreferences';
 export { ErrorUtils } from './utils/ErrorUtils';
 export { ShadowDomManager } from './utils/ShadowDomManager';
 // Localization